backend/requirements.txt
# Backend Python Dependencies - AtlasIQ Web

# =============================================================================
# Core Framework
# =============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# =============================================================================
# Database & ORM
# =============================================================================
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# =============================================================================
# Redis & Caching
# =============================================================================
redis[hiredis]==5.0.1
aioredis==2.0.1

# =============================================================================
# Authentication & Security
# =============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1

# =============================================================================
# HTTP & API Clients
# =============================================================================
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# =============================================================================
# Data Processing
# =============================================================================
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# =============================================================================
# Background Jobs & Scheduling
# =============================================================================
apscheduler==3.10.4
celery[redis]==5.3.4

# =============================================================================
# WebSocket Support
# =============================================================================
python-socketio==5.10.0
websockets==12.0

# =============================================================================
# Configuration & Environment
# =============================================================================
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# =============================================================================
# Validation & Serialization
# =============================================================================
email-validator==2.1.0
phonenumbers==8.13.26

# =============================================================================
# Monitoring & Logging
# =============================================================================
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# =============================================================================
# Testing
# =============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0

# =============================================================================
# Code Quality
# =============================================================================
black==23.12.0
ruff==0.1.8
mypy==1.7.1
pre-commit==3.6.0

# =============================================================================
# Documentation
# =============================================================================
mkdocs==1.5.3
mkdocs-material==9.5.2

# =============================================================================
# Utilities
# =============================================================================
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3
cachetools==5.3.2

# =============================================================================
# Development Tools
# =============================================================================
ipython==8.18.1
ipdb==0.13.13
watchdog==3.0.0

# =============================================================================
# Type Stubs
# =============================================================================
types-redis==4.6.0.11
types-requests==2.31.0.10
types-pytz==2023.3.1.1
